refactor objective logic
