+++ pypesto/objective/amici_objective.py
@@ -114,29 +114,6 @@ class AmiciObjective(Objective):
             # otherwise 1 can be guaranteed
             max_sensi_order = 2 if amici_model.o2mode else 1
 
-        fun = self.get_bound_fun()
-
-        if max_sensi_order > 0:
-            grad = True
-            hess = True
-        else:
-            grad = None
-            hess = None
-
-        res = self.get_bound_res()
-
-        if max_sensi_order > 0:
-            sres = True
-        else:
-            sres = None
-
-        super().__init__(
-            fun=fun, grad=grad, hess=hess, hessp=None,
-            res=res, sres=sres,
-            fun_accept_sensi_orders=True,
-            res_accept_sensi_orders=True,
-        )
-
         self.amici_model = amici.ModelPtr(amici_model.clone())
         self.amici_solver = amici.SolverPtr(amici_solver.clone())
 
@@ -190,7 +167,6 @@ class AmiciObjective(Objective):
         if x_names is None:
             # use ids as names
             x_names = x_ids
-        self.x_names = x_names
 
         self.n_threads = n_threads
         self.amici_object_builder = amici_object_builder
@@ -198,49 +174,12 @@ class AmiciObjective(Objective):
         if calculator is None:
             calculator = AmiciCalculator()
         self.calculator = calculator
+        super().__init__(x_names=x_names)
 
     def initialize(self):
         super().initialize()
         self.reset_steadystate_guesses()
         self.calculator.initialize()
-
-    def get_bound_fun(self):
-        """
-        Generate a fun function that calls _call_amici with MODE_FUN. Defining
-        a non-class function that references self as a local variable will bind
-        the function to a copy of the current self object and will
-        accordingly not take future changes to self into account.
-        """
-        def fun(x, sensi_orders):
-            return self._call_amici(x, sensi_orders, MODE_FUN)
-
-        return fun
-
-    def get_bound_res(self):
-        """
-        Generate a res function that calls _call_amici with MODE_RES. Defining
-        a non-class function that references self as a local variable will bind
-        the function to a copy of the current self object and will
-        accordingly not take future changes to self into account.
-        """
-        def res(x, sensi_orders):
-            return self._call_amici(x, sensi_orders, MODE_RES)
-
-        return res
-
-    def rebind_fun(self):
-        """
-        Replace the current fun function with one that is bound to the current
-        instance
-        """
-        self.fun = self.get_bound_fun()
-
-    def rebind_res(self):
-        """
-        Replace the current res function with one that is bound to the current
-        instance
-        """
-        self.res = self.get_bound_res()
 
     def __deepcopy__(self, memodict: Dict = None) -> 'AmiciObjective':
         other = self.__class__.__new__(self.__class__)
@@ -253,10 +192,6 @@ class AmiciObjective(Objective):
         other.amici_model = amici.ModelPtr(self.amici_model.clone())
         other.amici_solver = amici.SolverPtr(self.amici_solver.clone())
         other.edatas = [amici.ExpData(data) for data in self.edatas]
-
-        # rebind functions for __call__
-        other.rebind_fun()
-        other.rebind_res()
 
         return other
 
@@ -315,20 +250,35 @@ class AmiciObjective(Objective):
         self.amici_solver = solver
         self.edatas = edatas
 
-    def _call_amici(
+    def _check_sensi_orders(self, sensi_orders, mode) -> None:
+        sensi_order = self._get_amici_sensi_order(sensi_orders)
+
+        if self.max_sensi_order is None:
+            if mode == MODE_FUN:
+                max_sensi_order = 1 + self.amici_model.o2mode
+            else:
+                max_sensi_order = 1
+        else:
+            max_sensi_order = self.max_sensi_order
+
+        if sensi_order > max_sensi_order:
+            raise ValueError(f"Objective cannot be called with sensi_orders="
+                             f"{sensi_orders} and mode={mode}")
+
+    def _get_amici_sensi_order(self, sensi_orders: Tuple[int, ...]) -> int:
+        # amici is built such that only the maximum sensitivity is required,
+        # the lower orders are then automatically computed
+        # order 2 currently not implemented, we are using the FIM
+        return min(max(sensi_orders), 1)
+
+    def call_unprocessed(
             self,
             x: np.ndarray,
             sensi_orders: Tuple[int, ...],
             mode: str
     ) -> Dict:
-        # amici is built such that only the maximum sensitivity is required,
-        # the lower orders are then automatically computed
-        sensi_order = min(max(sensi_orders), 1)
-        # order 2 currently not implemented, we are using the FIM
 
-        # check if the requested sensitivities can be computed
-        if sensi_order > self.max_sensi_order:
-            raise Exception("Sensitivity order not allowed.")
+        sensi_order = self._get_amici_sensi_order(sensi_orders)
 
         x_dct = self.par_arr_to_dct(x)
