+++ pypesto/objective/objective.py
@@ -2,7 +2,7 @@ import numpy as np
 import pandas as pd
 import copy
 import logging
-from typing import Callable, Dict, Sequence, Tuple, Union
+from typing import Dict, Sequence, Tuple, Union
 
 from .constants import MODE_FUN, MODE_RES, FVAL, GRAD, HESS, RES, SRES
 from .history import HistoryBase
@@ -24,71 +24,6 @@ class Objective:
     by this objective function. If maximization is to be performed, the sign
     should be flipped before creating the objective function.
 
-    Parameters
-    ----------
-
-    fun:
-        The objective function to be minimized. If it only computes the
-        objective function value, it should be of the form
-
-            ``fun(x) -> float``
-
-        where x is an 1-D array with shape (n,), and n is the parameter space
-        dimension.
-
-    grad:
-        Method for computing the gradient vector. If it is a callable,
-        it should be of the form
-
-            ``grad(x) -> array_like, shape (n,).``
-
-        If its value is True, then fun should return the gradient as a second
-        output.
-
-    hess:
-        Method for computing the Hessian matrix. If it is a callable,
-        it should be of the form
-
-            ``hess(x) -> array, shape (n,n).``
-
-        If its value is True, then fun should return the gradient as a
-        second, and the Hessian as a third output, and grad should be True as
-        well.
-
-    hessp:
-        Method for computing the Hessian vector product, i.e.
-
-            ``hessp(x, v) -> array_like, shape (n,)``
-
-        computes the product H*v of the Hessian of fun at x with v.
-
-    res:
-        Method for computing residuals, i.e.
-
-            ``res(x) -> array_like, shape(m,).``
-
-    sres:
-        Method for computing residual sensitivities. If its is a callable,
-        it should be of the form
-
-            ``sres(x) -> array, shape (m,n).``
-
-        If its value is True, then res should return the residual
-        sensitivities as a second output.
-
-    fun_accept_sensi_orders:
-        Flag indicating whether fun takes sensi_orders as an argument.
-        Default: False.
-
-    res_accept_sensi_orders:
-        Flag indicating whether res takes sensi_orders as an argument.
-        Default: False
-
-    x_names:
-        Parameter names. None if no names provided, otherwise a list of str,
-        length dim_full (as in the Problem class). Can be read by the
-        problem.
-
     Attributes
     ----------
 
@@ -100,35 +35,10 @@ class Objective:
         Preprocess input values to and postprocess output values from
         __call__. Configured in `update_from_problem()`.
 
-    Notes
-    -----
-
-    If fun_accept_sensi_orders resp. res_accept_sensi_orders is True,
-    fun resp. res can also return dictionaries instead of tuples.
-    In that case, they are expected to follow the naming conventions
-    in ``constants.py``. This is of interest, because when __call__ is
-    called with return_dict = True, the full dictionary is returned, which
-    can contain e.g. also simulation data or debugging information.
     """
 
     def __init__(self,
-                 fun: Callable = None,
-                 grad: Union[Callable, bool] = None,
-                 hess: Callable = None,
-                 hessp: Callable = None,
-                 res: Callable = None,
-                 sres: Union[Callable, bool] = None,
-                 fun_accept_sensi_orders: bool = False,
-                 res_accept_sensi_orders: bool = False,
                  x_names: Sequence[str] = None):
-        self.fun = fun
-        self.grad = grad
-        self.hess = hess
-        self.hessp = hessp
-        self.res = res
-        self.sres = sres
-        self.fun_accept_sensi_orders = fun_accept_sensi_orders
-        self.res_accept_sensi_orders = res_accept_sensi_orders
 
         self.x_names = x_names
 
@@ -176,7 +86,7 @@ class Objective:
     def has_sres(self) -> bool:
         return callable(self.sres) or self.sres is True
 
-    def check_sensi_orders(self, sensi_orders, mode) -> None:
+    def _check_sensi_orders(self, sensi_orders, mode) -> None:
         """
         Check if the objective is able to compute the requested
         sensitivities. If not, throw an exception.
@@ -186,17 +96,20 @@ class Objective:
         ValueError if the objective function cannot be called as
         requested.
         """
-        if (mode is MODE_FUN and
-            (0 in sensi_orders and not self.has_fun
-             or 1 in sensi_orders and not self.has_grad
-             or 2 in sensi_orders and not self.has_hess)
-            ) or (mode is MODE_RES and
-                  (0 in sensi_orders and not self.has_res
-                   or 1 in sensi_orders and not self.has_sres)
-                  ):
-            raise ValueError(
-                f"Objective cannot be called with sensi_orders={sensi_orders}"
-                f" and mode={mode}")
+        raise NotImplementedError()
+
+    @staticmethod
+    def _check_mode(mode) -> None:
+        """
+        Check if the objective is able to compute in the requested mode
+
+        Raises
+        ------
+        ValueError if the objective function cannot be called as
+        requested.
+        """
+        if mode not in [MODE_FUN, MODE_RES]:
+            raise ValueError("This mode is not supported.")
 
     def __call__(
             self,
@@ -241,13 +154,14 @@ class Objective:
         x = np.array(x).copy()
 
         # check input
-        self.check_sensi_orders(sensi_orders, mode)
+        self._check_mode(mode)
+        self._check_sensi_orders(sensi_orders, mode)
 
         # pre-process
         x_full = self.pre_post_processor.preprocess(x)
 
         # compute result
-        result = self._call_unprocessed(x_full, sensi_orders, mode)
+        result = self.call_unprocessed(x_full, sensi_orders, mode)
 
         # post-process
         result = self.pre_post_processor.postprocess(result)
@@ -261,7 +175,7 @@ class Objective:
 
         return result
 
-    def _call_unprocessed(
+    def call_unprocessed(
             self,
             x: np.ndarray,
             sensi_orders: Tuple[int, ...],
@@ -269,148 +183,23 @@ class Objective:
     ) -> Dict:
         """
         Call objective function without pre- or post-processing and
-        formatting.
+        formatting. Abstract method to be implemented by derived classes.
+
+        Parameters
+        ----------
+        x:
+            The parameters for which to evaluate the objective function.
+        sensi_orders:
+            Specifies which sensitivities to compute, e.g. (0,1) -> fval, grad.
+        mode:
+            Whether to compute function values or residuals.
 
         Returns
         -------
         result:
             A dict containing the results.
         """
-        if mode == MODE_FUN:
-            result = self._call_mode_fun(x, sensi_orders)
-        elif mode == MODE_RES:
-            result = self._call_mode_res(x, sensi_orders)
-        else:
-            raise ValueError("This mode is not supported.")
-        return result
-
-    def _call_mode_fun(
-            self, x: np.ndarray, sensi_orders: Tuple[int, ...]
-    ) -> Dict:
-        """
-        The method __call__ was called with mode MODE_FUN.
-        """
-        if self.fun_accept_sensi_orders:
-            result = self.fun(x, sensi_orders)
-            if not isinstance(result, dict):
-                result = Objective.output_to_dict(
-                    sensi_orders, MODE_FUN, result)
-        elif sensi_orders == (0,):
-            if self.grad is True:
-                fval = self.fun(x)[0]
-            else:
-                fval = self.fun(x)
-            result = {FVAL: fval}
-        elif sensi_orders == (1,):
-            if self.grad is True:
-                grad = self.fun(x)[1]
-            else:
-                grad = self.grad(x)
-            result = {GRAD: grad}
-        elif sensi_orders == (2,):
-            if self.hess is True:
-                hess = self.fun(x)[2]
-            else:
-                hess = self.hess(x)
-            result = {HESS: hess}
-        elif sensi_orders == (0, 1):
-            if self.grad is True:
-                fval, grad = self.fun(x)[0:2]
-            else:
-                fval = self.fun(x)
-                grad = self.grad(x)
-            result = {FVAL: fval,
-                      GRAD: grad}
-        elif sensi_orders == (1, 2):
-            if self.hess is True:
-                grad, hess = self.fun(x)[1:3]
-            else:
-                hess = self.hess(x)
-                if self.grad is True:
-                    grad = self.fun(x)[1]
-                else:
-                    grad = self.grad(x)
-            result = {GRAD: grad,
-                      HESS: hess}
-        elif sensi_orders == (0, 1, 2):
-            if self.hess is True:
-                fval, grad, hess = self.fun(x)[0:3]
-            else:
-                hess = self.hess(x)
-                if self.grad is True:
-                    fval, grad = self.fun(x)[0:2]
-                else:
-                    fval = self.fun(x)
-                    grad = self.grad(x)
-            result = {FVAL: fval,
-                      GRAD: grad,
-                      HESS: hess}
-        else:
-            raise ValueError("These sensitivity orders are not supported.")
-        return result
-
-    def _call_mode_res(
-            self, x: np.ndarray, sensi_orders: Tuple[int, ...]
-    ) -> Dict:
-        """
-        The method __call__ was called with mode MODE_RES.
-        """
-        if self.res_accept_sensi_orders:
-            result = self.res(x, sensi_orders)
-            if not isinstance(result, dict):
-                result = Objective.output_to_dict(
-                    sensi_orders, MODE_RES, result)
-        elif sensi_orders == (0,):
-            if self.sres is True:
-                res = self.res(x)[0]
-            else:
-                res = self.res(x)
-            result = {RES: res}
-        elif sensi_orders == (1,):
-            if self.sres is True:
-                sres = self.res(x)[1]
-            else:
-                sres = self.sres(x)
-            result = {SRES: sres}
-        elif sensi_orders == (0, 1):
-            if self.sres is True:
-                res, sres = self.res(x)
-            else:
-                res = self.res(x)
-                sres = self.sres(x)
-            result = {RES: res,
-                      SRES: sres}
-        else:
-            raise ValueError("These sensitivity orders are not supported.")
-        return result
-
-    @staticmethod
-    def output_to_dict(
-            sensi_orders: Tuple[int, ...], mode: str, output_tuple: Tuple
-    ) -> Dict:
-        """
-        Convert output tuple to dict.
-        """
-        output_dict = {}
-        index = 0
-        if not isinstance(output_tuple, tuple):
-            output_tuple = (output_tuple,)
-        if mode == MODE_FUN:
-            if 0 in sensi_orders:
-                output_dict[FVAL] = output_tuple[index]
-                index += 1
-            if 1 in sensi_orders:
-                output_dict[GRAD] = output_tuple[index]
-                index += 1
-            if 2 in sensi_orders:
-                output_dict[HESS] = output_tuple[index]
-        elif mode == MODE_RES:
-            if 0 in sensi_orders:
-                output_dict[RES] = output_tuple[index]
-                index += 1
-            if 1 in sensi_orders:
-                output_dict[SRES] = output_tuple[index]
-        return output_dict
+        raise NotImplementedError()
 
     @staticmethod
     def output_to_tuple(
@@ -439,7 +228,6 @@ class Objective:
         return output
 
     # The following are convenience functions for getting specific outputs.
-
     def get_fval(self, x: np.ndarray) -> float:
         """
         Get the function value at x.
